requirements/development.txt
"""
 development requirements for {{ project_name }} project.
"""
# http://django-debug-toolbar.readthedocs.org/
django-debug-toolbar==1.3.2